+++ lib/iris/tests/unit/fileformats/netcdf/loader/ugrid_load/__init__.py
@@ -2,4 +2,4 @@
 #
 # This file is part of Iris and is released under the BSD license.
 # See LICENSE in the root of the repository for full licensing details.
-"""Unit tests for the :mod:`iris.mesh.load` package."""
+"""Unit tests for the :mod:`iris.fileformats.netcdf.ugrid_load` package."""
