+++ lib/iris/tests/unit/fileformats/netcdf/loader/ugrid_load/test_load_mesh.py
@@ -8,14 +8,15 @@
 # importing anything else.
 import iris.tests as tests  # isort:skip
 
-from iris.mesh.load import load_mesh
+from iris.mesh import load_mesh
 
 
 class Tests(tests.IrisTest):
     # All 'real' tests have been done for load_meshes(). Here we just check
     #  that load_mesh() works with load_meshes() correctly, using mocking.
     def setUp(self):
-        self.load_meshes_mock = self.patch("iris.mesh.load.load_meshes")
+        tgt = "iris.fileformats.netcdf.ugrid_load.load_meshes"
+        self.load_meshes_mock = self.patch(tgt)
         # The expected return from load_meshes - a dict of files, each with
         #  a list of meshes.
         self.load_meshes_mock.return_value = {"file": ["mesh"]}
